Feature/157774516 disable filter options with no results (#2)
* Typo

* Fix prop types in demo result elements

* Beef up documentation

* Externalise CSS styles for input
This is only necessary when integrating the component with a framework
that styles the `input` elements. It also avoids problems with CSS
modules down the line.

* Migrate to Webpack 4 with webpack-command and webpack-serve; update deps

* Replace inline styles with styled-components

* Update snapshots

* Remove prop to hide facet group name It’s not being used at all

* Refactor prop types used by several components

* Split main component in a stateful container and a stateless results list
Both the list and the sidebar get the results list and the selected
facets, from which they can infer what should and shouldn’t be displayed

* Rename demo pages and Webpack entries to reflect new code structure

* Fix tests

* Remove leftovers of facet group name hiding

* Revert: display results message D:

* Don’t wrap FacetPropTypes in PropTypes.shape to make it extendible

* Add a disabled prop to facets to grey out or remove the facet

* The container is now in charge of enabling/disabling facets
This design centralises all state changes in one place. I think it’s
better than having a distributed state stored among different
components.

* DRY out components: FilterList goes back to being as dumb as before

* Fix FilterList tests

* Fix facet group tests

* Randomize test

* Sort facets alphabetically; fix tests
The test suite includes a flakey test: the id in the input elements
created by react-select change on subsequent runs. Pushing to see if it
happens in Travis as well.

* Fix tests for FacetedSearchContainer

* Infer episode facets from episodes

* Properly randomize test

* Test that disabled checkboxes are displayed but greyed out

* Update dependencies

* Rename pass-through result component class prop

* Fix test after renaming props

* More fine-grained control of adding/removing facets for edge cases
We need to also know it a facet is the first selected or last unselected
in a group to, besides the general case which wasn’t being considered.

* Another easy test and updated snapshots